requirements.txt
# Scraping
selenium==4.18.1
webdriver-manager==4.0.1

# Data
pandas==2.2.1
numpy==1.26.4

# Base de données
sqlalchemy==2.0.28
pymysql==1.1.0
cryptography==42.0.5

# API
flask==3.0.2
flask-swagger-ui==4.11.1
flask-cors==4.0.0
prometheus-flask-exporter==0.23.0

# Celery + Redis
celery==5.3.6
redis==5.0.3
flower==2.0.1

# Config
python-dotenv==1.0.1

# Tests
pytest==8.1.1
pytest-cov==5.0.0
